Split SceneHierarchy and Inspector
